requirements.txt
fastapi[standard]
chromadb 
sentence-transformers
groq
python-dotenv
python-jose[cryptography] 
bcrypt==4.0.1
passlib[bcrypt]
sqlalchemy